start work on grid calculations in postgis

+++ 7. calculate contours/qgis_processing_script.py
@@ -10,7 +10,8 @@ from qgis.core import (QgsProcessing,
                        QgsProject,
                        QgsFieldConstraints,
                        QgsProcessingParameterVectorLayer,
-                       QgsProcessingFeedback)
+                       QgsProcessingFeedback,
+                       QgsCoordinateReferenceSystem)
 import processing
 import traceback
 import os
@@ -20,41 +21,41 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
     # used when calling the algorithm from another algorithm, or when
     # calling from the QGIS console.
 
-    INPUT = 'INPUT'
+    CHAINAGE = 'CHAINAGE'
     EQUIDISTANCE = 'EQUIDISTANCE'
-    CLIP = 'CLIP'
     OUTPUT = 'OUTPUT'
+    DATABASE = 'DATABASE'
 
     def tr(self, string):
-        """Returns a translatable string with the self.tr() function."""
+        '''Returns a translatable string with the self.tr() function.'''
         return QCoreApplication.translate('Processing', string)
 
     def createInstance(self):
         return ExampleProcessingAlgorithm()
 
     def name(self):
-        """Returns the algorithm name, used for identifying the algorithm."""
+        '''Returns the algorithm name, used for identifying the algorithm.'''
         return 'contour_batch'
 
     def displayName(self):
-        """Returns the translated algorithm name, which should be used for any user-visible display of the algorithm name."""
-        return self.tr('Contour Batch')
+        '''Returns the translated algorithm name, which should be used for any user-visible display of the algorithm name.'''
+        return self.tr('Grid preparation')
 
     def group(self):
-        """Returns the name of the group this algorithm belongs to. """
+        '''Returns the name of the group this algorithm belongs to. '''
         return self.tr('Lunis')
 
     def groupId(self):
-        """Returns the unique ID of the group this algorithm belongs to."""
+        '''Returns the unique ID of the group this algorithm belongs to.'''
         return 'lunis'
 
     def initAlgorithm(self, config=None):
-        """Here we define the inputs and output of the algorithm, along with some other properties."""
+        '''Here we define the inputs and output of the algorithm, along with some other properties.'''
         
         self.addParameter(
             QgsProcessingParameterVectorLayer(
-                self.INPUT,
-                self.tr('Input layer'),
+                self.CHAINAGE,
+                self.tr('Chainaged Network layer (used for filled contours)'),
                 [QgsProcessing.TypeVectorPoint]
             )
         )
@@ -62,24 +63,23 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
         self.addParameter(
             QgsProcessingParameterNumber(
                 self.EQUIDISTANCE,
-                self.tr('Equidistance (m)'),
+                self.tr('Equidistance of Contour Lines (m)'),
                 defaultValue=500,
                 minValue=1
             )
         )
         
         self.addParameter(
-            QgsProcessingParameterVectorLayer(
-                self.CLIP,
-                self.tr('Clip layer'),
-                [QgsProcessing.TypeVectorPolygon]
-            )
-        )
-    
-        self.addParameter(
             QgsProcessingParameterFolderDestination(
                 self.OUTPUT,
                 self.tr('Output folder'),
+            )
+        )
+        
+        self.addParameter(
+            QgsProcessingParameterString(
+                self.DATABASE,
+                self.tr('Database name'),
             )
         )
         
@@ -89,9 +89,9 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
         Here is where the processing itself takes place.
         '''
         
-        input_layer = self.parameterAsVectorLayer(
+        chainage_layer = self.parameterAsVectorLayer(
             parameters,
-            self.INPUT,
+            self.CHAINAGE,
             context
         )
         
@@ -101,20 +101,20 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
             context
         )
         
-        clip_layer = self.parameterAsVectorLayer(
-            parameters,
-            self.CLIP,
-            context
-        )
-        
         output_directory = self.parameterAsString(
             parameters,
             self.OUTPUT,
             context
         )
         
+        database = self.parameterAsString(
+            parameters,
+            self.DATABASE,
+            context
+        )
+        
         #get the count of all needed fields
-        count_of_school_fields = self.get_count_of_school_fields(input_layer)
+        count_of_school_fields = self.get_count_of_school_fields(chainage_layer)
         total = 100.0 / count_of_school_fields if count_of_school_fields > 0 else 0
         
         #init the result map
@@ -123,11 +123,15 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
         #init parts of the different file names
         prefix_contour = 'contour_'
         prefix_clip = 'clip_'
+        prefix_points = 'regular_points'
+        prefix_points_joined = 'points_joined_'
+        prefix_grid_joined = 'grid_joined_'
         suffix_polygon = '_polygon.geojson'
         suffix_linestring = '_linestring.geojson'
+        suffix_file_format = '.geojson'
         
         #iterate over all fields of this layer
-        for current, field in enumerate(input_layer.fields()):
+        for current, field in enumerate(chainage_layer.fields()):
             #print(str(current))
             # Stop the algorithm if cancel button has been clicked
             if feedback.isCanceled():
@@ -136,10 +140,10 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
             #evaluate the name of the current field
             if field.name().startswith("school_id"):
                 #add a filter expression to the current field
-                input_layer.setSubsetString(field.name() + " is not null")
+                chainage_layer.setSubsetString(field.name() + " is not null")
                 
                 #start the processing
-                count_of_contours = self.getNumberOfContours(input_layer, field.name(), equidistance)
+                count_of_contours = self.getNumberOfContours(chainage_layer, field.name(), equidistance)
                 
                 try:
                     #create polygons
@@ -151,7 +155,7 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
                                     'ContourType' : 1,
                                     'ExtendOption' : 0,
                                     'InputField' : field.name(),
-                                    'InputLayer' : input_layer,
+                                    'InputLayer' : chainage_layer,
                                     'LabelTrimZeros' : False,
                                     'LabelDecimalPlaces' : 0,
                                     'LabelUnits' : 'm',
@@ -160,77 +164,59 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
                                     'NContour' : 1000000,
                                     'OutputLayer' : QDir.toNativeSeparators(output_directory + '/' + prefix_contour + field.name() + suffix_polygon)})
                 
-                    result.update({'1 - ' + field.name() + ' - polygon': 'ok'})
+                    result.update({'1 - ' + field.name() + ' - filled contours': 'ok'})
                 except:
-                    result.update({'1 - ' + field.name() + ' - polygon': traceback.format_exc()})
+                    result.update({'1 - ' + field.name() + ' - filled contours': traceback.format_exc()})
                 
+                #import the new file/layer into a PostgreSQL-DB
                 try:
-                    #create linestrings
-                    processing.run('contourplugin:generatecontours',
-                                   {'ContourInterval' : equidistance,
-                                    'DuplicatePointTolerance': 0.0,
-                                    'ContourLevels' : '',
-                                    'ContourMethod' : 3,
-                                    'ContourType' : 0,
-                                    'ExtendOption' : 0,
-                                    'InputField' : field.name(),
-                                    'InputLayer' : input_layer,
-                                    'LabelTrimZeros' : False,
-                                    'LabelDecimalPlaces' : 0,
-                                    'LabelUnits' : 'm',
-                                    'MaxContourValue' : None,
-                                    'MinContourValue' : None,
-                                    'NContour' : 1000000,
-                                    'OutputLayer' : QDir.toNativeSeparators(output_directory + '/' + prefix_contour + field.name() + suffix_linestring)})
-                    
-                    result.update({'2 - ' + field.name() + ' - linestring': 'ok'})
-                except:
-                    result.update({'2 - ' + field.name() + ' - linestring': traceback.format_exc()})
-                
-                try:
-                    #clip the contours
-                    processing.run('qgis:clip',
+                    #create polygons
+                    processing.run('qgis:importintopostgis',
                                    {'INPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_contour + field.name() + suffix_polygon),
-                                    'OVERLAY' : clip_layer,
-                                    'OUTPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_clip + field.name() + suffix_polygon)})
-                    processing.run('qgis:clip',
-                                   {'INPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_contour + field.name() + suffix_linestring),
-                                    'OVERLAY' : clip_layer,
-                                    'OUTPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_clip + field.name() + suffix_linestring)})
-                    
-                    result.update({'3 - ' + field.name() + ' - clip': 'ok'})
-                except:
-                    result.update({'3 - ' + field.name() + ' - clip': traceback.format_exc()})
+                                    'DATABASE': database,
+                                    'SCHEMA' : 'grid_preparation',
+                                    'TABLENAME' : 'filled_contours',
+                                    'PRIMARY_KEY' : 'id',
+                                    'GEOMETRY_COLUMN' : 'geom',
+                                    'ENCODING' : 'UTF-8',
+                                    'OVERWRITE' : True,
+                                    'CREATEINDEX' : True,
+                                    'LOWERCASE_NAMES' : True,
+                                    'DROP_STRING_LENGTH' : True,
+                                    'FORCE_SINGLEPART' : True})
                 
-                try:
-                    #reproject the cliped layers into EPSG:4326
-                    processing.run('native:reprojectlayer',
-                                   {'INPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_clip + field.name() + suffix_polygon),
-                                    'TARGET_CRS' : 'EPSG:4326',
-                                    'OUTPUT' : QDir.toNativeSeparators(output_directory + '/' + field.name() + suffix_polygon)})
-                    processing.run('native:reprojectlayer',
-                                   {'INPUT' : QDir.toNativeSeparators(output_directory + '/' + prefix_clip + field.name() + suffix_linestring),
-                                    'TARGET_CRS' : 'EPSG:4326',
-                                    'OUTPUT' : QDir.toNativeSeparators(output_directory + '/' + field.name() + suffix_linestring)})
-                    
-                    result.update({'4 - ' + field.name() + ' - reprojection': 'ok'})
+                    result.update({'2 - ' + field.name() + ' - database import': 'ok'})
                 except:
-                    result.update({'4 - ' + field.name() + ' - reprojection': traceback.format_exc()})
-
-                #remove the filter expression to the current field
-                input_layer.setSubsetString("")
+                    result.update({'2 - ' + field.name() + ' - database import': traceback.format_exc()})
+                
+                #calculate new grid values
+                try:
+                    sql_regular_points = ""
+                    
+                    processing.run('qgis:postgisexecutesql',
+                                   {'DATABASE' : database,
+                                    'SQL' : sql_regular_points})
+                
+                    result.update({'3 - ' + field.name() + ' - SQL to calculate grid values': 'ok'})
+                except:
+                    result.update({'3 - ' + field.name() + ' - SQL to calculate grid values': traceback.format_exc()})
+                
+                #remove the filter expression of the current field
+                chainage_layer.setSubsetString("")
                 
                 #update progressbar
                 feedback.setProgress(int(current * total))
                 
         #remove all temporary files
-        self.removeTempFiles(output_directory, prefix_contour)
-        self.removeTempFiles(output_directory, prefix_clip)
+        self.removeTempFiles(output_directory, prefix_contour, suffix_polygon)
+        #self.removeTempFiles(output_directory, prefix_points, suffix_file_format)
+        
+        #load the grid
         
         return result
     
     def get_count_of_school_fields(self, layer):
-        '''This function iterates over all fields of a given layer and returns the count of fields starting with "school".'''
+        '''This function iterates over all fields of a given layer and returns the count of fields starting with *school*.'''
         count = 0
         for field in layer.fields():
             if field.name().startswith("school"):
@@ -249,10 +235,10 @@ class ExampleProcessingAlgorithm(QgsProcessingAlgorithm):
         
         return number_of_contours
     
-    def removeTempFiles(self, directory, prefix):
+    def removeTempFiles(self, directory, prefix, suffix):
         '''This function removes all temporary files from the given directory.'''
         for entry in os.listdir(directory):
             absolute_path = QDir.toNativeSeparators(directory + '/' + entry)
             if os.path.isfile(absolute_path):
-                if os.path.basename(entry).startswith(prefix):
+                if os.path.basename(entry).startswith(prefix) and os.path.basename(entry).endswith(suffix):
                     os.remove(absolute_path)
